Jump logic moved
jump logic moved to control stage in order not to lose too many
expression, modified the register file in order to read asynchronously.
Data ram changed so that there is a read enable to avoid wrong
addresses. branch and jump tested succesfully and jal not tested yet.
Ready to implement forwarding.